+++ python/iqm/qdmi/__init__.py
@@ -20,7 +20,12 @@
 from ._paths import IQM_QDMI_CMAKE_DIR, IQM_QDMI_INCLUDE_DIR, IQM_QDMI_LIBRARY_PATH
 from ._version import version as __version__
 
-__all__ = ["IQM_QDMI_CMAKE_DIR", "IQM_QDMI_INCLUDE_DIR", "IQM_QDMI_LIBRARY_PATH", "__version__"]
+__all__ = [
+    "IQM_QDMI_CMAKE_DIR",
+    "IQM_QDMI_INCLUDE_DIR",
+    "IQM_QDMI_LIBRARY_PATH",
+    "__version__",
+]
 
 
 def __dir__() -> list[str]:
@@ -35,6 +40,8 @@ try:
     from qiskit.primitives.containers.data_bin import DataBin
 
     pickle.loads(pickle.dumps(DataBin()))  # noqa: S301
+except ImportError:
+    pass
 except NotImplementedError:
     # Bypass the immutable __setattr__ restriction during unpickling
     DataBin.__setattr__ = object.__setattr__